Attempt to fix Sagemaker docker test
