+++ mlflow/utils/file_utils.py
@@ -228,36 +228,3 @@ def write_to(filename, data):
 def append_to(filename, data):
     with open(filename, "a") as handle:
         handle.write(data)
-
-
-def _copy_mlflow_project(mlflow_path, output_dir=""):
-    """
-    Internal function used to copy mlflow project during development.
-    The mlflow is assumed to be accessible as a local directory in this case.
-
-    :param output_dir: mlflow will be copied here
-    :return: name of the mlflow project directory
-    """
-    def _docker_ignore(mlflow_root):
-        docker_ignore = os.path.join(mlflow_root, '.dockerignore')
-        patterns = []
-        if os.path.exists(docker_ignore):
-            with open(docker_ignore, "r") as f:
-                patterns = [x.strip() for x in f.readlines()]
-
-        def ignore(_, names):
-            import fnmatch
-            res = set()
-            for p in patterns:
-                res.update(set(fnmatch.filter(names, p)))
-            return list(res)
-
-        return ignore if patterns else None
-
-    mlflow_root = os.path.dirname(mlflow_path)
-    if not os.path.exists(mlflow_root):
-        raise Exception("Can not find local mlflow copy at '%s'" % mlflow_root)
-    mlflow_dir = os.path.basename(mlflow_root)
-    shutil.copytree(mlflow_root, os.path.join(output_dir, mlflow_dir),
-                    ignore=_docker_ignore(mlflow_root))
-    return mlflow_dir
